Create index.js
index.js includes the show/hide functions for the createBookPost modal, as well as the search/filter functions, the creation/insertion of new book posts made, and the event listeners for all these interactions. 

One important thing to note is that there are parameters for seller information (which would be provided by the account the seller has logged in on as), but I'm not sure how we plan on linking that up so I've just inserted placeholder values for those parameters for now.